Configure eslint pretter
