+++ configs/eslint-next.js
@@ -15,7 +15,7 @@ module.exports = {
     'plugin:import/typescript',
     'prettier',
   ],
-  plugins: ['import', '@typescript-eslint', 'unused-imports'],
+  plugins: ['import', '@typescript-eslint', 'unused-imports', 'prettier'],
   settings: {
     next: {
       rootDir: [
@@ -55,6 +55,7 @@ module.exports = {
     'import/prefer-default-export': 'off',
     'no-param-reassign': ['error', { props: false }],
     'no-console': 'warn',
+    "prettier/prettier": ["error"],
     'react/require-default-props': 0,
     'react/no-array-index-key': 0,
     'react/jsx-props-no-spreading': 0,
