+++ package.json
@@ -75,6 +75,7 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jest": "^26.1.1",
+    "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-testing-library": "^5.0.1",
     "eslint-plugin-unused-imports": "^2.0.0",
     "husky": "^7.0.4",
